ID DataType changed from BigInteger to Long

+++ src/main/java/com/example/personal_budget/dto/CreateTransactionRequest.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 
 @Data
 public class CreateTransactionRequest {
-  private BigInteger userID;
+  private Long userID;
 
   @Positive(message = "Amount must be positive")
   private double amount;
@@ -22,7 +22,7 @@ public class CreateTransactionRequest {
   @JsonFormat(pattern = "yyyy-MM-dd")
   private LocalDate date;
 
-  private BigInteger categoryID;
+  private Long categoryID;
   private String source;
   private String description;
 }
