+++ src/main/java/com/example/personal_budget/entity/Transaction.java
@@ -18,11 +18,11 @@ public class Transaction {
 
     @Id
     @GeneratedValue
-    BigInteger id;
+    Long id;
 
     //@ManyToOne
     //@JoinColumn(name = "user_id")
-    BigInteger userID;
+    Long userID;
     double amount;
 
     @Enumerated(EnumType.STRING)
@@ -30,7 +30,7 @@ public class Transaction {
 
     // @ManyToOne
     // @JoinColumn(name = "category_id")
-    BigInteger categoryID;
+    Long categoryID;
     LocalDate date;
     String source;
     String description;
